adding code that was accidentally deleted in an earlier merge
