+++ client/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
+    "react-redux": "^5.0.6",
     "react-scripts": "1.0.11"
   },
   "scripts": {
